Integrate norway (#20)
* add skeleton for norway integration

* add skeleton for norway integration

* rename NobilPipeline.py and integrate sweden

* fix location

* add sweden

* ignore exported json files in git

* remove unused parameter

* update readme

---------
