+++ .gitignore
@@ -5,5 +5,6 @@ venv/
 data/
 .python-version
 stations_*.csv
+stations_*.geo.json
 testdata_merge.csv
 token_deepatlas.json
