+++ charging_stations_pipelines/pipelines/ocm/ocm_extractor.py
@@ -70,6 +70,14 @@ def testSth(x):
 
 
 def ocm_extractor(tmp_file_path: str, country_code: str):
+
+    # OCM export contains norwegian data under country code "NO" and that's why we need to rename it to "NO"
+    if country_code == "NOR":
+        country_code = "NO"
+    if country_code == "SWE":
+        country_code = "SE"
+
+
     project_data_dir: str = pathlib.Path(tmp_file_path).parent.resolve()
     data_root_dir: str = os.path.join(project_data_dir, "ocm-export")
     data_dir: str = os.path.join(data_root_dir, f"data/{country_code}")
