+++ charging_stations_pipelines/pipelines/shared.py
@@ -1,7 +1,9 @@
 import logging
+from dateutil import parser
 from numbers import Number
 
 logger = logging.getLogger(__name__)
+
 
 def check_coordinates(coords: float) -> float:
     if isinstance(coords, str):
@@ -12,3 +14,11 @@ def check_coordinates(coords: float) -> float:
     if not isinstance(coords, Number):
         raise ValueError("Coordinatess could not be read propery!")
     return coords
+
+
+def parse_date(date):
+    try:
+        return parser.parse(date)
+    except TypeError as e:
+        logger.debug(f"Could not parse DateCreated! {e}")
+        return None
